Created core module and integrated with data and domain

+++ data/src/main/java/com/weatherapp/data/entities/WeatherDaoEntity.kt
@@ -8,12 +8,12 @@ import com.google.gson.annotations.SerializedName
 data class WeatherDaoEntity(
     @PrimaryKey(autoGenerate = true) val cod: Long,
     @SerializedName("name") val name: String,
-    @SerializedName("id")  val id: Long,
-    @SerializedName("temp") val temp: Long,
-    @SerializedName("feels_like") val feels_like: Long,
-    @SerializedName("temp_min") val temp_min: Long,
-    @SerializedName("temp_max") val temp_max: Long,
-    @SerializedName("pressure") val pressure: Long,
-    @SerializedName("humidity") val humidity: Long,
-    @SerializedName("lon") val lon: Long,
-    @SerializedName("lat") val lat: Long)
+    @SerializedName("id")  val id: Float,
+    @SerializedName("temp") val temp: Float,
+    @SerializedName("feels_like") val feels_like: Float,
+    @SerializedName("temp_min") val temp_min: Float,
+    @SerializedName("temp_max") val temp_max: Float,
+    @SerializedName("pressure") val pressure: Float,
+    @SerializedName("humidity") val humidity: Float,
+    @SerializedName("lon") val lon: Float,
+    @SerializedName("lat") val lat: Float)
